+++ src/lib.rs
@@ -15,5 +15,6 @@ pub mod output;
 pub mod parser;
 pub mod quantity;
 pub mod runtime;
+pub mod text_runtime;
 
 pub const VERSION: &str = env!("CARGO_PKG_VERSION");